docker-compose config, renamed java package. resolves #4
