+++ pom.xml
@@ -81,7 +81,7 @@
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>org.example.Main</mainClass>
+                            <mainClass>de.hsh.capstoneris.Main</mainClass>
                         </manifest>
                     </archive>
                     <descriptorRefs>
@@ -111,7 +111,7 @@
                     </execution>
                 </executions>
                 <configuration>
-                    <mainClass>org.example.Main</mainClass>
+                    <mainClass>de.hsh.capstoneris.Main</mainClass>
                 </configuration>
             </plugin>
         </plugins>
