+++ src/test/java/de/hsh/capstoneris/CustomerTest.java
@@ -1,4 +1,4 @@
-package org.example;
+package de.hsh.capstoneris;
 
 import javax.ws.rs.client.Client;
 import javax.ws.rs.client.ClientBuilder;
@@ -8,7 +8,7 @@ import org.glassfish.grizzly.http.server.HttpServer;
 
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Test;
+
 import static org.junit.Assert.assertEquals;
 
 public class CustomerTest {
